+++ pulumi-language-dotnet/testdata/projects/l2-resource-primitives/Program.cs
@@ -13,9 +13,9 @@ return await Deployment.RunAsync(() =>
         String = "hello",
         NumberArray = new[]
         {
-            (double)-1,
-            (double)0,
-            (double)1,
+            -1.0,
+            0.0,
+            1.0,
         },
         BooleanMap = 
         {
